Add ability to configure gateway ip from args
